requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
autopep8==2.0.4
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
pycodestyle==2.11.0
PyJWT==2.8.0
python-logstash==0.4.8
pytz==2023.3
redis==5.0.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
